MediateDataGenerationParameters

Convenience class for use with other methods and classes to generate mediation data for use with parallel processing.

Slots

n

is the sample size.

pX

is the minor allele frequency

gamma0

is the intercept for M

gammaX

is the association of X with M

varM

is the variance of M

beta0

is the intercept for Y

betaX

is the direct effect of X on Y

betaM

is a vector of different associations of M with Y

varY

is the variance of Y

nSim

is the number of simulations to run